Rename legacy catalina package to org.jboss.modcluster.container.tomcat
