+++ container/tomcat/src/test/java/org/jboss/modcluster/container/tomcat/AutoProxyConnectorProviderTestCase.java
@@ -1,4 +1,25 @@
-package org.jboss.modcluster.container.catalina;
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2017, Red Hat Middleware LLC, and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package org.jboss.modcluster.container.tomcat;
 
 import static org.junit.Assert.assertSame;
 import static org.mockito.ArgumentMatchers.same;
@@ -8,8 +29,6 @@ import static org.mockito.Mockito.when;
 import org.apache.catalina.Engine;
 import org.apache.catalina.Service;
 import org.jboss.modcluster.container.Connector;
-import org.jboss.modcluster.container.tomcat.AutoProxyConnectorProvider;
-import org.jboss.modcluster.container.tomcat.ConnectorFactory;
 import org.junit.Test;
 
 public class AutoProxyConnectorProviderTestCase {
