+++ container/tomcat/src/test/java/org/jboss/modcluster/container/tomcat/ServiceLoaderTomcatFactoryTestCase.java
@@ -1,4 +1,25 @@
-package org.jboss.modcluster.container.catalina;
+/*
+ * JBoss, Home of Professional Open Source.
+ * Copyright 2017, Red Hat Middleware LLC, and individual contributors
+ * as indicated by the @author tags. See the copyright.txt file in the
+ * distribution for a full listing of individual contributors.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+ */
+package org.jboss.modcluster.container.tomcat;
 
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
@@ -7,19 +28,6 @@ import org.jboss.modcluster.container.Context;
 import org.jboss.modcluster.container.Engine;
 import org.jboss.modcluster.container.Host;
 import org.jboss.modcluster.container.Server;
-import org.jboss.modcluster.container.tomcat.TomcatConnectorFactory;
-import org.jboss.modcluster.container.tomcat.TomcatContextFactory;
-import org.jboss.modcluster.container.tomcat.TomcatEngineFactory;
-import org.jboss.modcluster.container.tomcat.TomcatFactoryRegistry;
-import org.jboss.modcluster.container.tomcat.TomcatHostFactory;
-import org.jboss.modcluster.container.tomcat.TomcatServerFactory;
-import org.jboss.modcluster.container.tomcat.ConnectorFactory;
-import org.jboss.modcluster.container.tomcat.ContextFactory;
-import org.jboss.modcluster.container.tomcat.EngineFactory;
-import org.jboss.modcluster.container.tomcat.HostFactory;
-import org.jboss.modcluster.container.tomcat.ProxyConnectorProvider;
-import org.jboss.modcluster.container.tomcat.ServerFactory;
-import org.jboss.modcluster.container.tomcat.ServiceLoaderTomcatFactory;
 import org.junit.Test;
 
 public class ServiceLoaderTomcatFactoryTestCase {
