+++ container/tomcat8/src/test/java/org/jboss/modcluster/container/tomcat8/ServiceLoaderTomcatFactoryTestCase.java
@@ -33,7 +33,7 @@ import org.jboss.modcluster.container.tomcat.TomcatHostFactory;
 /**
  * @author Paul Ferraro
  */
-public class ServiceLoaderTomcatFactoryTestCase extends org.jboss.modcluster.container.catalina.ServiceLoaderTomcatFactoryTestCase {
+public class ServiceLoaderTomcatFactoryTestCase extends org.jboss.modcluster.container.tomcat.ServiceLoaderTomcatFactoryTestCase {
     @Override
     protected void verifyCatalinaFactoryTypes(TomcatFactoryRegistry registry) {
         assertSame(registry.getServerFactory().getClass(), TomcatServerFactory.class);
