+++ container/tomcat85/src/test/java/org/jboss/modcluster/container/tomcat85/EngineTestCase.java
@@ -24,5 +24,5 @@ package org.jboss.modcluster.container.tomcat85;
 /**
  * @author Paul Ferraro
  */
-public class EngineTestCase extends org.jboss.modcluster.container.catalina.EngineTestCase {
+public class EngineTestCase extends org.jboss.modcluster.container.tomcat.EngineTestCase {
 }
